Tests/LibWeb/Text/expected/video-canceled-load.txt

file:///example/file/location/wfh
